ShellPkg: Fix ARM build errors.
submitted-by: Olivier Martin olivier.martin@arm.com>

git-svn-id: https://svn.code.sf.net/p/edk2/code/trunk/edk2@14400 6f19259b-4bc3-4df7-8a09-765794883524

+++ ShellPkg/Library/UefiShellLevel2CommandsLib/TimeDate.c
@@ -667,7 +667,7 @@ CheckAndSetTimeZone (
   if (*Walker == L'-') {
     TheTime.TimeZone = (INT16)((ShellStrToUintn (++Walker)) * 60);
   } else {
-    TheTime.TimeZone = (INT16)((ShellStrToUintn (Walker)) * -60);
+    TheTime.TimeZone = (INT16)((INT16)(ShellStrToUintn (Walker)) * -60);
   }
   if (Walker2 != NULL) {
     Walker = Walker2 + 1;
